+++ CHANGELOG.md
@@ -4,9 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.8] / 2022-12-05
+### Features
+- Feature `ValidateTestAssemblyNUnitVersion` - NUnit version `3.13.3.0`
+### Added
+- Add `ValidateTestAssemblyNUnitVersion`
+- Add Command/TestUtils
+
 ## [1.0.7] / 2022-12-01
 ### Features
-- Featere `TestEngine.GetTestFullNames` supported
+- Feature `TestEngine.GetTestFullNames` supported
 - Feature `WildcardPattern` filter supported
 - Feature `TestCaseAttribute` supported
 ### Added
@@ -75,6 +82,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Test with Console Output
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.0.8]: ../../compare/1.0.7...1.0.8
 [1.0.7]: ../../compare/1.0.6...1.0.7
 [1.0.6]: ../../compare/1.0.5...1.0.6
 [1.0.5]: ../../compare/1.0.4...1.0.5